Start migrating verify invite endpoint
